+++ src/hooks/useAudioEngine.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react"
+import { AudioEngineContext } from "../components";
+
+export const useAudioEngine = () => {
+  return useContext(AudioEngineContext);
+};